Add RemoveStarColor
This change now allows the custom control to highlight the stars that will be removed when the mouse goes before the current selected star, highlighted by the RemoveColorStar property. To account for this change, OnMouseMove will now once again start from 0, but due to the additional condition, the HoverStarColor will still begin from the last selected star.